+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.08] - 2024-11-26
+
+### Added
+
+- Add Input component
+- Add Styling to signup
+- Add progress bar to signup
+- Add loading animation to signup
+- Add pretty create animation to signup
+
 ## [0.2.07] - 2024-11-24
 
 - Add Step component